requirements.txt
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
croniter==2.0.5
tenacity==8.5.0
